+++ header/src/main/java/org/zstack/header/vm/VmAttachOtherDiskExtensionPoint.java
@@ -3,6 +3,6 @@ package org.zstack.header.vm;
 import org.zstack.header.core.Completion;
 
 public interface VmAttachOtherDiskExtensionPoint {
-    void attachOtherDiskToVm(APICreateVmInstanceMsg.DiskAO diskAO, String vmInstanceUuid, Completion completion);
+    void attachOtherDiskToVm(DiskAO diskAO, String vmInstanceUuid, Completion completion);
     String getDiskType();
 }
