+++ header/src/main/java/org/zstack/header/vm/VmInstanceSpec.java
@@ -397,13 +397,13 @@ public class VmInstanceSpec implements Serializable {
     }
 
     private List<String> disableL3Networks;
-    private List<APICreateVmInstanceMsg.DiskAO> diskAOs;
+    private List<DiskAO> diskAOs;
 
-    public List<APICreateVmInstanceMsg.DiskAO> getDiskAOs() {
+    public List<DiskAO> getDiskAOs() {
         return diskAOs;
     }
 
-    public void setDiskAOs(List<APICreateVmInstanceMsg.DiskAO> diskAOs) {
+    public void setDiskAOs(List<DiskAO> diskAOs) {
         this.diskAOs = diskAOs;
     }
